chore: moved store/utils/error.ts file to utils/error.ts
