+++ src/widgets/BhandaraCheckin/store/utils/index.ts
@@ -1,4 +1,2 @@
-export * from "./errors";
-
 export const getBhandaraCheckinActionName = (name: string): string =>
   `bhandara-checkin/${name}`;
