+++ src/widgets/BhandaraCheckin/utils/index.ts
@@ -1,1 +1,2 @@
 export * from "./validations";
+export * from "./errors";
